Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaDevices-enumerateDevices.https-expected.txt

Description

This test checks for the presence of the navigator.mediaDevices.enumerateDevices() method.


PASS mediaDevices.enumerateDevices() is present and working 
FAIL InputDeviceInfo is supported promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: InputDeviceInfo"
